Refactor the code in the backend to make it cleaner
