+++ backend/controllers/project.controller.js
@@ -66,57 +66,4 @@ ProjectController.destroy = async function (req, res) {
   }
 };
 
-const { User } = require('../models');
-
-async function getProjectById(projectId) {
-  if (projectId === '64077da036505ba8e73810b2') return null;
-  try {
-    const project = await Project.findById(projectId);
-    if (!project) {
-      console.warn(`Project not found for ID: ${projectId}`);
-      return null; // Ensure it returns null if not found
-    }
-    return project;
-  } catch (error) {
-    console.error('Error fetching project:', error);
-    return null; // Return null on error to avoid breaking the app
-  }
-}
-
-ProjectController.getProjectManagers = async function (req, res) {
-  try {
-    // const userProjectMap = {};
-    const userProjectMap = [];
-    const projectManagers = await User.find({
-      $and: [
-        { accessLevel: { $in: ['admin', 'superadmin'] } },
-        { managedProjects: { $exists: true, $type: 'array', $ne: [] } },
-      ],
-    });
-
-    for (const projectManager of projectManagers) {
-      projectManager.accessLevel = 'projectLead';
-
-      const projectNames = [];
-
-      for (const projectId of projectManager.managedProjects) {
-        // console.log('project', typeof project);
-        const projectDetail = await getProjectById(projectId);
-        if (projectDetail) {
-          projectNames.push(projectDetail.name);
-          // console.log('projectDetailName', projectDetail.name);
-        } else {
-          console.warn('Project detail is null, cannot access name');
-        }
-      }
-      // console.log('projectNames', projectNames);
-      projectManager.managedProjects = projectNames;
-    }
-    console.log('projectManager', projectManagers[0]);
-    return res.status(200).send(projectManagers);
-  } catch (err) {
-    return res.sendStatus(400);
-  }
-};
-
 module.exports = ProjectController;
