+++ backend/routers/projects.router.js
@@ -1,13 +1,10 @@
-const express = require("express");
+const express = require('express');
 const router = express.Router();
 
 const { ProjectController } = require('../controllers');
-const { AuthUtil } = require("../middleware");
+const { AuthUtil } = require('../middleware');
 
 // The base is /api/projects
-
-router.get('/projectManagers', ProjectController.getProjectManagers);
-
 router.get('/', ProjectController.project_list);
 
 // Its a put because we have to send the PM projects to be filtered here
@@ -20,6 +17,5 @@ router.get('/:ProjectId', ProjectController.project_by_id);
 router.put('/:ProjectId', AuthUtil.verifyCookie, ProjectController.update);
 
 router.patch('/:ProjectId', AuthUtil.verifyCookie, ProjectController.update);
-
 
 module.exports = router;
