+++ client/src/api/UserApiService.js
@@ -36,6 +36,20 @@ class UserApiService {
     return [];
   }
 
+  async fetchProjectsManagers() {
+    try {
+      const route = this.baseUserUrl + '/projectManagers';
+      const res = await fetch(route, {
+        headers: this.headers,
+      });
+      return await res.json();
+    } catch (error) {
+      console.error(`fetchProjectsManagers error: ${error}`);
+      alert('Server not responding. Please refresh the page.');
+    }
+    return [];
+  }
+
   // Updates user projects in db
   async updateUserDbProjects(userToEdit, managedProjects) {
     // eslint-disable-next-line no-underscore-dangle
